Small refactor before merge
