+++ app/src/main/java/com/example/roomer/presentation/screens/shared_screens/user_details_screen/UserDetailsScreen.kt
@@ -20,6 +20,7 @@ import androidx.compose.material.FloatingActionButton
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -44,10 +45,10 @@ import com.example.roomer.domain.model.entities.User
 import com.example.roomer.domain.model.login_sign_up.InterestModel
 import com.example.roomer.presentation.screens.destinations.ChatScreenDestination
 import com.example.roomer.presentation.screens.entrance.signup.habits_screen.HabitTileModel
+import com.example.roomer.presentation.screens.shared_screens.user_details_screen.FollowButton
 import com.example.roomer.presentation.screens.shared_screens.user_details_screen.UserDetailsScreenViewModel
 import com.example.roomer.presentation.ui_components.BackBtn
 import com.example.roomer.presentation.ui_components.ExpandableText
-import com.example.roomer.presentation.ui_components.FollowButton
 import com.example.roomer.presentation.ui_components.GreenButtonOutline
 import com.example.roomer.presentation.ui_components.GreenButtonOutlineIconed
 import com.example.roomer.presentation.ui_components.HabitsTable
@@ -67,6 +68,7 @@ fun UserDetailsScreen(
     viewModel: UserDetailsScreenViewModel = hiltViewModel()
 ) {
     NavbarManagement.hideNavbar()
+    val screenState = viewModel.state.collectAsState().value
     val textStyleHeadline = TextStyle(
         color = colorResource(id = R.color.black),
         fontWeight = FontWeight.Medium,
@@ -110,7 +112,7 @@ fun UserDetailsScreen(
             ) {
                 UserAvatar(avatarUrl = user.avatar)
                 FollowButton(
-                    followManipulate = viewModel.followManipulate,
+                    isFollowed = screenState.isFollow,
                     followUserId = user.userId,
                     currentUserId = viewModel.currentUser.value.userId
                 )
